third_party/blink/web_tests/http/tests/devtools/sources/debugger-ui/show-generator-location-expected.txt

Tests that "Show Generator Location" jumps to the correct location.


Running: testiterNotStarted
Generator location revealed: [20:20]

Running: testiterSuspended1
Generator location revealed: [22:11]

Running: testiterSuspended2
Generator location revealed: [23:11]

Running: testiterSuspended3
Generator location revealed: [24:11]

Running: testiterClosed
Generator location revealed: [20:20]
